+++ samples/test/transcoder.test.js
@@ -34,13 +34,29 @@ const preset = 'preset/web-hd';
 const templateName = `projects/${projectNumber}/locations/${location}/jobTemplates/${templateId}`;
 
 const testFileName = 'ChromeCast.mp4';
+const testOverlayFileName = 'overlay.jpg';
+
 const inputUri = `gs://${bucketName}/${testFileName}`;
+const overlayUri = `gs://${bucketName}/${testOverlayFileName}`;
 const outputUriForPreset = `gs://${bucketName}/test-output-preset/`;
 const outputUriForTemplate = `gs://${bucketName}/test-output-template/`;
 const outputUriForAdHoc = `gs://${bucketName}/test-output-adhoc/`;
+const outputUriForStaticOverlay = `gs://${bucketName}/test-output-static-overlay/`;
+const outputUriForAnimatedOverlay = `gs://${bucketName}/test-output-animated-overlay/`;
+const outputDirForSetNumberImagesSpritesheet =
+  'test-output-set-number-spritesheet/';
+const outputUriForSetNumberImagesSpritesheet = `gs://${bucketName}/${outputDirForSetNumberImagesSpritesheet}`;
+const outputDirForPeriodicImagesSpritesheet =
+  'test-output-periodic-spritesheet/';
+const outputUriForPeriodicImagesSpritesheet = `gs://${bucketName}/${outputDirForPeriodicImagesSpritesheet}`;
+// Spritesheets are generated from the input video into the bucket directories above.
+// Spritesheets use the following file naming conventions:
+const smallSpriteSheetFileName = 'small-sprite-sheet0000000000.jpeg';
+const largeSpriteSheetFileName = 'large-sprite-sheet0000000000.jpeg';
 
 const cwd = path.join(__dirname, '..');
-const resourceFile = `testdata/${testFileName}`;
+const videoFile = `testdata/${testFileName}`;
+const overlayFile = `testdata/${testOverlayFileName}`;
 
 function wait(ms) {
   return new Promise(resolve => {
@@ -50,6 +66,16 @@ function wait(ms) {
   });
 }
 
+const checkFileExists = async function (bucketName, fileName) {
+  const [files] = await storage.bucket(bucketName).getFiles();
+  for (let i = 0; i < files.length; i++) {
+    if (files[i].name === fileName) {
+      return true;
+    }
+  }
+  return false;
+};
+
 before(async () => {
   assert(
     process.env.GOOGLE_CLOUD_PROJECT_NUMBER,
@@ -57,7 +83,8 @@ before(async () => {
   );
   // Create a Cloud Storage bucket to be used for testing.
   await storage.createBucket(bucketName);
-  await storage.bucket(bucketName).upload(resourceFile);
+  await storage.bucket(bucketName).upload(videoFile);
+  await storage.bucket(bucketName).upload(overlayFile);
 });
 
 after(async () => {
@@ -157,13 +184,11 @@ describe('Job functions preset', () => {
 
   it('should check that the job succeeded', async function () {
     this.retries(5);
-    await wait(30000);
+    await wait(90000);
     const output = execSync(
       `node getJobState.js ${projectId} ${location} ${this.presetJobId}`,
       {cwd}
     );
-    // TODO(bcoe): remove this debug information once passing:
-    console.info(output.toString('utf8'));
     assert.ok(output.includes('Job state: SUCCEEDED'));
   });
 });
@@ -217,13 +242,11 @@ describe('Job functions template', () => {
 
   it('should check that the job succeeded', async function () {
     this.retries(5);
-    await wait(30000);
+    await wait(90000);
     const output = execSync(
       `node getJobState.js ${projectId} ${location} ${this.templateJobId}`,
       {cwd}
     );
-    // TODO(bcoe): remove this debug information once passing:
-    console.info(output.toString('utf8'));
     assert.ok(output.includes('Job state: SUCCEEDED'));
   });
 });
@@ -267,13 +290,205 @@ describe('Job functions adhoc', () => {
 
   it('should check that the job succeeded', async function () {
     this.retries(5);
-    await wait(30000);
+    await wait(90000);
     const output = execSync(
       `node getJobState.js ${projectId} ${location} ${this.adhocJobId}`,
       {cwd}
     );
-    // TODO(bcoe): remove this debug information once passing:
-    console.info(output.toString('utf8'));
     assert.ok(output.includes('Job state: SUCCEEDED'));
+  });
+});
+
+describe('Job with static overlay functions', () => {
+  before(function () {
+    const output = execSync(
+      `node createJobWithStaticOverlay.js ${projectId} ${location} ${inputUri} ${overlayUri} ${outputUriForStaticOverlay}`,
+      {cwd}
+    );
+    assert.ok(
+      output.includes(`projects/${projectNumber}/locations/${location}/jobs/`)
+    );
+    this.staticOverlayJobId = output.toString().split('/').pop();
+  });
+
+  after(function () {
+    const output = execSync(
+      `node deleteJob.js ${projectId} ${location} ${this.staticOverlayJobId}`,
+      {cwd}
+    );
+    assert.ok(output.includes('Deleted job'));
+  });
+
+  it('should get a job', function () {
+    const output = execSync(
+      `node getJob.js ${projectId} ${location} ${this.staticOverlayJobId}`,
+      {cwd}
+    );
+    const jobName = `projects/${projectNumber}/locations/${location}/jobs/${this.staticOverlayJobId}`;
+    assert.ok(output.includes(jobName));
+  });
+
+  it('should check that the job succeeded', async function () {
+    this.retries(5);
+    await wait(90000);
+    const output = execSync(
+      `node getJobState.js ${projectId} ${location} ${this.staticOverlayJobId}`,
+      {cwd}
+    );
+    assert.ok(output.includes('Job state: SUCCEEDED'));
+  });
+});
+
+describe('Job with animated overlay functions', () => {
+  before(function () {
+    const output = execSync(
+      `node createJobWithAnimatedOverlay.js ${projectId} ${location} ${inputUri} ${overlayUri} ${outputUriForAnimatedOverlay}`,
+      {cwd}
+    );
+    assert.ok(
+      output.includes(`projects/${projectNumber}/locations/${location}/jobs/`)
+    );
+    this.animatedOverlayJobId = output.toString().split('/').pop();
+  });
+
+  after(function () {
+    const output = execSync(
+      `node deleteJob.js ${projectId} ${location} ${this.animatedOverlayJobId}`,
+      {cwd}
+    );
+    assert.ok(output.includes('Deleted job'));
+  });
+
+  it('should get a job', function () {
+    const output = execSync(
+      `node getJob.js ${projectId} ${location} ${this.animatedOverlayJobId}`,
+      {cwd}
+    );
+    const jobName = `projects/${projectNumber}/locations/${location}/jobs/${this.animatedOverlayJobId}`;
+    assert.ok(output.includes(jobName));
+  });
+
+  it('should check that the job succeeded', async function () {
+    this.retries(5);
+    await wait(90000);
+    const output = execSync(
+      `node getJobState.js ${projectId} ${location} ${this.animatedOverlayJobId}`,
+      {cwd}
+    );
+    assert.ok(output.includes('Job state: SUCCEEDED'));
+  });
+});
+
+describe('Job with set number of images spritesheet', () => {
+  before(function () {
+    const output = execSync(
+      `node createJobWithSetNumberImagesSpritesheet.js ${projectId} ${location} ${inputUri} ${outputUriForSetNumberImagesSpritesheet}`,
+      {cwd}
+    );
+    assert.ok(
+      output.includes(`projects/${projectNumber}/locations/${location}/jobs/`)
+    );
+    this.setNumberSpritesheetJobId = output.toString().split('/').pop();
+  });
+
+  after(function () {
+    const output = execSync(
+      `node deleteJob.js ${projectId} ${location} ${this.setNumberSpritesheetJobId}`,
+      {cwd}
+    );
+    assert.ok(output.includes('Deleted job'));
+  });
+
+  it('should get a job', function () {
+    const output = execSync(
+      `node getJob.js ${projectId} ${location} ${this.setNumberSpritesheetJobId}`,
+      {cwd}
+    );
+    const jobName = `projects/${projectNumber}/locations/${location}/jobs/${this.setNumberSpritesheetJobId}`;
+    assert.ok(output.includes(jobName));
+  });
+
+  it('should check that the job succeeded', async function () {
+    this.retries(5);
+    await wait(90000);
+    const output = execSync(
+      `node getJobState.js ${projectId} ${location} ${this.setNumberSpritesheetJobId}`,
+      {cwd}
+    );
+    assert.ok(output.includes('Job state: SUCCEEDED'));
+  });
+
+  it('should check that the spritesheet files exist in the bucket', async () => {
+    assert.equal(
+      await checkFileExists(
+        bucketName,
+        `${outputDirForSetNumberImagesSpritesheet}${smallSpriteSheetFileName}`
+      ),
+      true
+    );
+    assert.equal(
+      await checkFileExists(
+        bucketName,
+        `${outputDirForSetNumberImagesSpritesheet}${largeSpriteSheetFileName}`
+      ),
+      true
+    );
+  });
+});
+
+describe('Job with periodic images spritesheet', () => {
+  before(function () {
+    const output = execSync(
+      `node createJobWithPeriodicImagesSpritesheet.js ${projectId} ${location} ${inputUri} ${outputUriForPeriodicImagesSpritesheet}`,
+      {cwd}
+    );
+    assert.ok(
+      output.includes(`projects/${projectNumber}/locations/${location}/jobs/`)
+    );
+    this.periodicSpritesheetJobId = output.toString().split('/').pop();
+  });
+
+  after(function () {
+    const output = execSync(
+      `node deleteJob.js ${projectId} ${location} ${this.periodicSpritesheetJobId}`,
+      {cwd}
+    );
+    assert.ok(output.includes('Deleted job'));
+  });
+
+  it('should get a job', function () {
+    const output = execSync(
+      `node getJob.js ${projectId} ${location} ${this.periodicSpritesheetJobId}`,
+      {cwd}
+    );
+    const jobName = `projects/${projectNumber}/locations/${location}/jobs/${this.periodicSpritesheetJobId}`;
+    assert.ok(output.includes(jobName));
+  });
+
+  it('should check that the job succeeded', async function () {
+    this.retries(5);
+    await wait(90000);
+    const output = execSync(
+      `node getJobState.js ${projectId} ${location} ${this.periodicSpritesheetJobId}`,
+      {cwd}
+    );
+    assert.ok(output.includes('Job state: SUCCEEDED'));
+  });
+
+  it('should check that the spritesheet files exist in the bucket', async () => {
+    assert.equal(
+      await checkFileExists(
+        bucketName,
+        `${outputDirForPeriodicImagesSpritesheet}${smallSpriteSheetFileName}`
+      ),
+      true
+    );
+    assert.equal(
+      await checkFileExists(
+        bucketName,
+        `${outputDirForPeriodicImagesSpritesheet}${largeSpriteSheetFileName}`
+      ),
+      true
+    );
   });
 });